requirements.txt
# FastAPI 및 서버 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 데이터베이스
asyncpg==0.29.0

# 데이터 검증 및 설정
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9

# RSS 수집
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
urllib3==2.1.0

# 중복 제거
simhash==2.1.2

# NLP (선택사항 - 현재는 규칙 기반만 사용)
# scikit-learn==1.3.2
# numpy==1.24.3
# pandas==2.1.3

# Redis (선택사항 - 필요시 Upstash 사용)
# redis==5.0.1

# Google OAuth 2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1



